+++ R/reexport.R
@@ -23,3 +23,21 @@ arkhe::remove_constant
 #' @importFrom arkhe sparsity
 #' @export
 arkhe::sparsity
+
+# REEXPORT FROM KHROMA
+
+#' @importFrom khroma color
+#' @export
+khroma::color
+
+#' @importFrom khroma palette_color_continuous
+#' @export
+khroma::palette_color_continuous
+
+#' @importFrom khroma palette_color_discrete
+#' @export
+khroma::palette_color_discrete
+
+#' @importFrom khroma palette_color_picker
+#' @export
+khroma::palette_color_picker